requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
black==24.10.0
boto3==1.35.41
botocore==1.35.41
celery==5.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.5
django-cors-headers==4.0.0
django-ninja==1.3.0
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
flake8==7.1.1
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.4.2
loguru==0.7.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.2
openai==0.27.2
packaging==24.1
pathspec==0.12.1
pinecone==5.3.1
pinecone-client==2.2.2
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.7
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
redis==4.5.1
requests==2.31.0
requests-mock==1.12.1
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.3